HM Portfolio | Android App Development
The Tech Portfolio App is an Android application designed to showcase your technical skills in a visually appealing and interactive format. It features a personalized introduction, skill-specific CardViews for areas like Programming, Android Development, and AI/ML, and detailed pages displaying the technologies you're proficient in. The app includes built-in email and phone call functionalities for easy communication, making it an ideal tool for networking, job applications, or client interactions. This project is fully customizable, offering a professional platform to highlight your expertise and achievements.